+++ scripts/mod_00_x=Omega_v_v=ams.py
@@ -59,6 +59,6 @@ params = {
 }
 
 # get average entanglement
-looper = wrap_looper(system=Mod00, params=params, func='ams', looper='x_looper', file_path='data/mod_00/entan_ln', plot=True)
+looper = wrap_looper(SystemClass=Mod00, params=params, func='ams', looper='x_looper', file_path='data/mod_00/entan_ln', plot=True)
 # calculate thresholds
 print(looper.get_thresholds(thres_mode='minmax'))